Revert "Extract code for converting a JS callback into an AWS function into its own location."
This reverts commit 85ae2b74d8a88a5174d431a70cf1af98a3831762.
