+++ resources.go
@@ -1268,7 +1268,6 @@ func Provider() tfbridge.ProviderInfo {
 			Files: []string{
 				"arn.ts",    // ARN typedef
 				"region.ts", // Region union type and constants
-				"utils.ts",  // General helper code
 			},
 			Modules: map[string]*tfbridge.OverlayInfo{
 				"config": {
@@ -1294,8 +1293,7 @@ func Provider() tfbridge.ProviderInfo {
 				},
 				"lambda": {
 					Files: []string{
-						"runtimes.ts",       // a union type and constants for available Lambda runtimes.
-						"createFunction.ts", // helper method to simply create a lambda function from an arrow function
+						"runtimes.ts", // a union type and constants for available Lambda runtimes.
 					},
 				},
 				"serverless": {
